Renaming CosmosDBProvider to be CosmosDBService
