--- NCS.DSS.GDPRUpdateCustomerInformation/Cosmos/Provider/ICosmosDBProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace NCS.DSS.GDPRUpdateCustomerInformation.Cosmos.Provider
-{
-    public interface ICosmosDBProvider
-    {
-        Task DeleteRecordsForCustomer(Guid customerId);
-    }
-}
