+++ NCS.DSS.GDPRUpdateCustomerInformation/Services/ICosmosDBService.cs
@@ -1,0 +1,7 @@
+﻿namespace NCS.DSS.GDPRUpdateCustomerInformation.Services
+{
+    public interface ICosmosDBService
+    {
+        Task DeleteRecordsForCustomer(Guid customerId);
+    }
+}
