+++ NCS.DSS.GDPRUpdateCustomerInformation/Services/IIdentifyAndAnonymiseDataService.cs
@@ -1,4 +1,4 @@
-﻿namespace NCS.DSS.GDPRUpdateCustomerInformation.Service
+﻿namespace NCS.DSS.GDPRUpdateCustomerInformation.Services
 {
     public interface IIdentifyAndAnonymiseDataService
     {
